+++ examples/test_example_check_check.py
@@ -1,0 +1,17 @@
+"""
+"with check.check:" - bad
+"with check:" - good
+
+However, we want both to work.
+"""
+
+
+def test_check(check):
+    with check:
+        assert True
+
+
+def test_check_check(check):
+    "Deprecated, but should still work for now"
+    with check.check:
+        assert True